Review fixes of address book's base
