+++ features/account/customer_account/address_book/viewing_addresses.feature
@@ -1,8 +1,8 @@
-@customer_account
-Feature: Viewing customer's addresses
-    In order to see all my addresses in address book
+@address_book
+Feature: Viewing my address book
+    In order to see all my addresses
     As a Customer
-    I want to be able to browse my addresses
+    I want to be able to browse my address book
 
     Background:
         Given the store operates on a single channel in "United States"
@@ -10,21 +10,21 @@ Feature: Viewing customer's addresses
 
     @ui
     Scenario: Viewing all addresses
-        Given I have an address "Lucifer Morningstar", "Seaside Fwy", "90802", "Los Angeles", "United States" in address book
-        When I browse my addresses
-        Then I should see a single address in the list
+        Given I have an address "Lucifer Morningstar", "Seaside Fwy", "90802", "Los Angeles", "United States", "Arkansas" in my address book
+        When I browse my address book
+        Then I should see a single address in my book
 
     @ui
     Scenario: Seeing only my addresses
         Given there is a customer "John Doe" identified by an email "doe@example.com" and a password "banana"
-        And this customer has an address "John Doe", "Banana Street", "90232", "New York", "United States" in address book
-        And I have an address "Lucifer Morningstar", "Seaside Fwy", "90802", "Los Angeles", "United States" in address book
-        When I browse my addresses
-        Then I should see a single address in the list
+        And this customer has an address "John Doe", "Banana Street", "90232", "New York", "United States", "Kansas" in their address book
+        And I have an address "Lucifer Morningstar", "Seaside Fwy", "90802", "Los Angeles", "United States", "Arkansas" in my address book
+        When I browse my address book
+        Then I should see a single address in my book
         And this address should be assigned to "Lucifer Morningstar"
-        And I should not see an address assigned to "John Doe"
+        And I should not see the address assigned to "John Doe"
 
     @ui
     Scenario: Viewing empty address book
-        When I browse my addresses
-        Then I should see information about no existing addresses
+        When I browse my address book
+        Then There should be no addresses
