+++ src/Sylius/Behat/Context/Ui/Shop/AccountContext.php
@@ -14,8 +14,6 @@ namespace Sylius\Behat\Context\Ui\Shop;
 use Behat\Behat\Context\Context;
 use Sylius\Behat\NotificationType;
 use Sylius\Behat\Page\PageInterface;
-use Sylius\Behat\Page\Shop\Account\AddressBook\CreatePageInterface as AddressBookCreatePageInterface;
-use Sylius\Behat\Page\Shop\Account\AddressBook\IndexPageInterface as AddressBookIndexPageInterface;
 use Sylius\Behat\Page\Shop\Account\ChangePasswordPageInterface;
 use Sylius\Behat\Page\Shop\Account\DashboardPageInterface;
 use Sylius\Behat\Page\Shop\Account\Order\IndexPageInterface;
@@ -23,7 +21,6 @@ use Sylius\Behat\Page\Shop\Account\Order\ShowPageInterface;
 use Sylius\Behat\Page\Shop\Account\ProfileUpdatePageInterface;
 use Sylius\Behat\Service\NotificationCheckerInterface;
 use Sylius\Component\Core\Formatter\StringInflector;
-use Sylius\Component\Core\Model\AddressInterface;
 use Sylius\Component\Core\Model\OrderInterface;
 use Webmozart\Assert\Assert;
 
@@ -58,16 +55,6 @@ final class AccountContext implements Context
     private $orderShowPage;
 
     /**
-     * @var AddressBookIndexPageInterface
-     */
-    private $addressBookIndexPage;
-
-    /**
-     * @var AddressBookCreatePageInterface
-     */
-    private $addressBookCreatePage;
-
-    /**
      * @var NotificationCheckerInterface
      */
     private $notificationChecker;
@@ -78,8 +65,6 @@ final class AccountContext implements Context
      * @param ChangePasswordPageInterface $changePasswordPage
      * @param IndexPageInterface $orderIndexPage
      * @param ShowPageInterface $orderShowPage
-     * @param AddressBookIndexPageInterface $addressBookIndexPage
-     * @param AddressBookCreatePageInterface $addressBookCreatePage
      * @param NotificationCheckerInterface $notificationChecker
      */
     public function __construct(
@@ -88,8 +73,6 @@ final class AccountContext implements Context
         ChangePasswordPageInterface $changePasswordPage,
         IndexPageInterface $orderIndexPage,
         ShowPageInterface $orderShowPage,
-        AddressBookIndexPageInterface $addressBookIndexPage,
-        AddressBookCreatePageInterface $addressBookCreatePage,
         NotificationCheckerInterface $notificationChecker
     ) {
         $this->dashboardPage = $dashboardPage;
@@ -97,8 +80,6 @@ final class AccountContext implements Context
         $this->changePasswordPage = $changePasswordPage;
         $this->orderIndexPage = $orderIndexPage;
         $this->orderShowPage = $orderShowPage;
-        $this->addressBookIndexPage = $addressBookIndexPage;
-        $this->addressBookCreatePage = $addressBookCreatePage;
         $this->notificationChecker = $notificationChecker;
     }
 
@@ -484,96 +465,6 @@ final class AccountContext implements Context
             $this->orderShowPage->hasBillingProvinceName($provinceName),
             sprintf('Cannot find shipping address with province %s', $provinceName)
         );
-    }
-
-    /**
-     * @When I browse my addresses
-     */
-    public function iBrowseMyAddresses()
-    {
-        $this->addressBookIndexPage->open();
-    }
-
-    /**
-     * @When I delete the :fullName address
-     */
-    public function iDeleteTheAddress($fullname)
-    {
-        $this->addressBookIndexPage->deleteAddress($fullname);
-    }
-
-    /**
-     * @Then I should be notified that it has been successfully deleted
-     */
-    public function iShouldBeNotifiedAboutSuccessfulDelete()
-    {
-        $this->notificationChecker->checkNotification('Address has been successfully deleted.', NotificationType::success());
-    }
-
-    /**
-     * @Then I should see a single address in the list
-     */
-    public function iShouldSeeASingleAddressInTheList()
-    {
-        Assert::true(
-            $this->addressBookIndexPage->isSingleAddressOnList(),
-            'There should be one address on the list, but it does not.'
-        );
-    }
-
-    /**
-     * @Then this address should be assigned to :fullName
-     * @Then the address assigned to :fullName should appear on the list
-     */
-    public function thisAddressShouldHavePersonFirstNameAndLastName($fullName)
-    {
-        Assert::true(
-            $this->addressBookIndexPage->hasAddressFullName($fullName),
-            sprintf('The full name in address should be %s, but it does not.', $fullName)
-        );
-    }
-
-    /**
-     * @Then I should see information about no existing addresses
-     */
-    public function iShouldSeeInformationAboutNoExistingAddresses()
-    {
-        Assert::true(
-            $this->addressBookIndexPage->hasNoExistingAddressesMessage(),
-            'There should be information about no existing addresses, but it does not.'
-        );
-    }
-
-    /**
-     * @Given I want to add a new address to my address book
-     */
-    public function iWantToAddANewAddressToMyAddressBook()
-    {
-        $this->addressBookCreatePage->open();
-    }
-
-    /**
-     * @When /^I specify its (data as "([^"]+)", "([^"]+)", "([^"]+)", "([^"]+)", "([^"]+)")$/
-     */
-    public function iSpecifyItsDataAs(AddressInterface $address)
-    {
-        $this->addressBookCreatePage->fillAddressData($address);
-    }
-
-    /**
-     * @Then I should be notified that it has been successfully added
-     */
-    public function iShouldBeNotifiedThatAdddressHasBeenSuccessfullyAdded()
-    {
-        $this->notificationChecker->checkNotification('has been successfully added.', NotificationType::success());
-    }
-
-    /**
-     * @When I add it
-     */
-    public function iAddIt()
-    {
-        $this->addressBookCreatePage->saveAddress();
     }
 
     /**
