+++ src/Sylius/Behat/Resources/config/suites.yml
@@ -12,6 +12,7 @@ imports:
     - suites/domain/promotion/managing_promotion_coupons.yml
 
     - suites/ui/account/customer.yml
+    - suites/ui/account/address_book.yml
     - suites/ui/account/login.yml
     - suites/ui/account/registration.yml
     - suites/ui/account/verification.yml
